requirement.txt
asgiref==3.7.2
astroid==3.1.0
autopep8==2.0.4
backports.zoneinfo==0.2.1
dill==0.3.8
Django==4.2.10
djangorestframework==3.14.0
isort==5.13.2
mccabe==0.7.0
pep8==1.7.1
platformdirs==4.2.0
pycodestyle==2.11.1
pylint==3.1.0
pytz==2024.1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.4
typing-extensions==4.10.0
